refactor: move `recv_manager` to `server/managers/recv_manager` for better management
